Moved class webserver tester do test\webserver namespace.
