+++ classes/test/webserver/webserver_tester.php
@@ -21,7 +21,7 @@
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace local_cleanurls;
+namespace local_cleanurls\test\webserver;
 
 use stdClass;
