+++ settings.php
@@ -44,7 +44,7 @@ if ($hassiteconfig) {
 
         // If we are on the settings page then also run a router test.
         if ($section == 'local_cleanurls') {
-            $tester = new \local_cleanurls\webserver_tester();
+            $tester = new \local_cleanurls\test\webserver\webserver_tester();
             if ($tester->test()) {
                 $test .= $OUTPUT->notification(get_string('routerok', 'local_cleanurls'), 'notifysuccess');
             } else {
